requirements.txt
# Core Dependencies
numpy>=1.24.0
pandas>=2.0.0
PyYAML>=6.0.1

# Machine Learning and NLP
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
chromadb>=0.4.0
networkx>=3.0
spacy>=3.8.0
regex>=2022.10.0
langdetect>=1.0.9

# PDF Processing
PyMuPDF>=1.21.0  # fitz kütüphanesi için

# Web API
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
python-multipart>=0.0.5  # Form data ve dosya yüklemeleri için

# Memory and Performance
psutil>=5.9.0

# Utilities
tqdm>=4.65.0
loguru>=0.7.0
python-dotenv>=1.0.0
pathlib>=1.0.1
huggingface-hub>=0.15.0
Pillow>=10.0.0

# Build ve Geliştirme Paketleri
pytest>=7.0.0
black>=23.3.0 